moved signal to utils
